refactor: 불필요한 추상화계층(Application) 제거
도메인 ID의 생성책임을 DB에 맡김
