--- src/main/java/cart/application/IdSequencer.java
@@ -1,10 +1,0 @@
-package cart.application;
-
-public class IdSequencer {
-
-    private static int sequence = 0;
-
-    public static int get() {
-        return ++sequence;
-    }
-}
