+++ src/main/java/cart/presentation/ProductController.java
@@ -1,7 +1,7 @@
 package cart.presentation;
 
 
-import cart.application.ProductCRUDApplication;
+import cart.business.ProductCRUDService;
 import cart.presentation.dto.ProductDto;
 import cart.presentation.dto.ProductIdDto;
 import org.springframework.http.ResponseEntity;
@@ -9,35 +9,40 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping(path = "/product")
 public class ProductController {
 
-    private final ProductCRUDApplication productCRUDApplication;
+    private final ProductCRUDService productCRUDService;
 
-    public ProductController(ProductCRUDApplication productCRUDApplication) {
-        this.productCRUDApplication = productCRUDApplication;
+    public ProductController(ProductCRUDService productCRUDService) {
+        this.productCRUDService = productCRUDService;
     }
 
     @PostMapping
     public void productCreate(@RequestBody @Valid ProductDto request) {
-        productCRUDApplication.create(request);
+        productCRUDService.create(ProductConverter.toProductWithoutId(request));
     }
 
     @GetMapping
     public ResponseEntity<List<ProductDto>> productRead() {
-        List<ProductDto> response = productCRUDApplication.readAll();
+        List<ProductDto> response = productCRUDService.readAll()
+                .stream()
+                .map(ProductConverter::toProductDto)
+                .collect(Collectors.toList());
+
         return ResponseEntity.ok(response);
     }
 
     @PutMapping
     public void productUpdate(@RequestBody @Valid ProductDto request) {
-        productCRUDApplication.update(request);
+        productCRUDService.update(ProductConverter.toProductWithId(request));
     }
 
     @DeleteMapping
     public void productDelete(@RequestBody @Valid ProductIdDto request) {
-        productCRUDApplication.delete(request);
+        productCRUDService.delete(request.getId());
     }
 }
