+++ src/main/java/cart/presentation/TemplateController.java
@@ -1,7 +1,7 @@
 package cart.presentation;
 
-import cart.application.ProductCRUDApplication;
 import cart.business.MemberReadService;
+import cart.business.ProductCRUDService;
 import cart.business.domain.member.Member;
 import cart.presentation.dto.ProductDto;
 import org.springframework.stereotype.Controller;
@@ -9,28 +9,37 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Controller
 public class TemplateController {
 
-    private final ProductCRUDApplication productCRUDApplication;
+    private final ProductCRUDService productCRUDService;
     private final MemberReadService memberReadService;
 
-    public TemplateController(ProductCRUDApplication productCRUDApplication, MemberReadService memberReadService) {
-        this.productCRUDApplication = productCRUDApplication;
+    public TemplateController(ProductCRUDService productCRUDService, MemberReadService memberReadService) {
+        this.productCRUDService = productCRUDService;
         this.memberReadService = memberReadService;
     }
 
     @GetMapping("/")
     public String home(Model model) {
-        List<ProductDto> products = productCRUDApplication.readAll();
+        List<ProductDto> products = productCRUDService.readAll()
+                .stream()
+                .map(ProductConverter::toProductDto)
+                .collect(Collectors.toList());
+
         model.addAttribute("products", products);
         return "index";
     }
 
     @GetMapping("/admin")
     public String admin(Model model) {
-        List<ProductDto> products = productCRUDApplication.readAll();
+        List<ProductDto> products = productCRUDService.readAll()
+                .stream()
+                .map(ProductConverter::toProductDto)
+                .collect(Collectors.toList());
+
         model.addAttribute("products", products);
         return "admin";
     }
